Switch to direct npm usage
The Makefile and it's usage of docker simply did not work,
after trying to make it work I decided to let it go and
switch to the classic npm usage.
I figured that if that did matter for the project, (which it
does not at the moment) I could switch back anytime later on.