Dug cells display the number of neighbor trapped cells
That could have been done within the Grid
(Domain get trappedNeighborByCoordinate + Component call to
function instead of cell property)
so that the information was kept out of the cells.
But I figured that this way the job was done only once and that
there was no real issue doing it that way.